Common modules added, some functions moved to modules: - created some handler files that are imported as modules   - modules have init functions preset in case of need and to make sure ide import clean wont clear module imports with only window functions - moved some functions to modules from common - data file rename to xHandler format - all modules use init function, even when just placeholder
